move InsertDeleteStep to util
